chore: count() 반환형 Integer로 변경

+++ src/main/java/in/koreatech/koin/domain/admin/land/repository/AdminLandRepository.java
@@ -10,7 +10,7 @@ public interface AdminLandRepository extends Repository<Land, Integer> {
 
     Page<Land> findAllByIsDeleted(boolean isDeleted, Pageable pageable);
 
-    Long countAllByIsDeleted(boolean isDeleted);
+    Integer countAllByIsDeleted(boolean isDeleted);
 
     Land save(Land request);
